requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
stripe==10.12.0
python-dotenv==1.0.1
requests==2.32.3

streamlit==1.39.0
pandas==2.2.3
openpyxl==3.1.5
